# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Demo Locker API",
    "version": "1.0.0",
    "description": "API for Demo Locker, a private-by-default demo/playlist sharing service. Most routes require a Bearer session token; media elements (<audio>/<img>) that cannot send an Authorization header instead accept the same token via a `?token=` query parameter (session token OR a per-playlist share/invite token — both are treated identically). Gated legacy routes never distinguish 'exists but private' from 'does not exist': any access failure returns a generic `{\"error\":\"not found\"}` 404 so private playlist/track/comment IDs cannot be enumerated by probing.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment (Worker or self-hosted)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token returned by POST /auth/signup or POST /auth/login, sent as `Authorization: Bearer <token>`."
      },
      "tokenQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Session token OR a share/invite token, sent as a `?token=` query parameter. Used by media elements (<audio src>, <img src>) that cannot set an Authorization header. Accepted anywhere bearerAuth is accepted on the legacy (non-/public/v1) gated routes."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "ThreadedComment": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Comment"
          },
          {
            "type": "object",
            "properties": {
              "replies": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/playlists": {
      "get": {
        "summary": "List the authenticated user's playlists",
        "operationId": "listPlaylists",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Playlists owned by the current user, ordered by createdAt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playlists": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Playlist"
                      }
                    }
                  },
                  "required": [
                    "playlists"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a playlist",
        "operationId": "createPlaylist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Playlist created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playlist": {
                      "$ref": "#/components/schemas/Playlist"
                    }
                  },
                  "required": [
                    "playlist"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Free-tier playlist limit reached (MAX_PLAYLISTS).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/{id}": {
      "get": {
        "summary": "Get a playlist and its tracks",
        "description": "Gated: requires a session acting in the playlist's locker (owner or collaborator) OR a valid share/invite token for this playlist, via Bearer header or `?token=`. A private/nonexistent/inaccessible playlist all return the same non-enumerable 404. Because this route admits anonymous share holders as well as locker sessions, neither the playlist's created_by nor any track's uploaded_by is serialized; createdByMe and uploadedByMe are computed from the caller's session instead — both false when there is none. The same applies to the attribution NAMES (createdByName, uploadedByName): a caller with no locker session is served null for every one of them — share links go to people outside the locker, and the band's names are not part of what a listen link grants.",
        "operationId": "getPlaylist",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "tokenQuery": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Session or share token (alternative to Authorization: Bearer)."
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist with its tracks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playlist": {
                      "$ref": "#/components/schemas/Playlist"
                    },
                    "tracks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Track"
                      }
                    }
                  },
                  "required": [
                    "playlist",
                    "tracks"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found — returned for private, nonexistent, or inaccessible playlists alike (non-enumerable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a playlist",
        "description": "Requires a session in the playlist's locker: the owner or one of their collaborators may rename it. Accepts partial updates to name, artworkKey, and isPublic (toggles public read access via /public/v1). isPublic is the locker owner's to change — a collaborator changing it in either direction is refused with 403.",
        "operationId": "updatePlaylist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "artworkKey": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "isPublic": {
                    "type": "boolean",
                    "description": "When true, the playlist becomes readable anonymously via /public/v1/playlists/{id}."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated playlist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playlist": {
                      "$ref": "#/components/schemas/Playlist"
                    }
                  },
                  "required": [
                    "playlist"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "isPublic is the locker owner's field to change — a collaborator attempting to publish or un-publish is refused. Unlike the 404 below, this is deliberately not non-enumerable: the caller can already read this playlist, so hiding its existence would only mislead them.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Outside the caller's locker, or does not exist (non-enumerable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a playlist",
        "description": "Locker-scoped, then narrowed by who created the playlist. The session must act in the playlist's locker, otherwise 404. Within the locker: the owner may delete any playlist; a collaborator may delete only one whose created_by is their own user id, and a null created_by reads as the owner's. Deleting a playlist does not delete its tracks — the FK is ON DELETE SET NULL, so they are detached back into the locker's library.",
        "operationId": "deletePlaylist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Non-enumerable. Three causes, indistinguishable to the caller: the playlist does not exist; it belongs to another locker; or it is in the caller's locker but the caller is a collaborator who did not create it (including any playlist with a null created_by).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/{id}/reorder": {
      "patch": {
        "summary": "Reorder a playlist's tracks",
        "description": "Locker-scoped: the playlist's locker owner or a collaborator session, or an `edit` share token for this playlist (Bearer header or `?token=`, no session required). Sets each track's position to its index in trackIds.",
        "operationId": "reorderPlaylistTracks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "trackIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                },
                "required": [
                  "trackIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reordered.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "trackIds is not an array.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not owned by the caller, or does not exist (non-enumerable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/{id}/artwork": {
      "get": {
        "summary": "Stream playlist artwork",
        "description": "Gated the same way as GET /playlists/{id}: owner session or a valid share token, via Bearer or `?token=` (an <img> tag cannot send an Authorization header). Private/nonexistent/no-artwork all 404 identically (non-enumerable).",
        "operationId": "getPlaylistArtwork",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "tokenQuery": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artwork image bytes. Cache-Control: private, max-age=3600.",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not found — private, nonexistent, or no artwork set (non-enumerable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload playlist artwork",
        "description": "Artwork must be a PNG, JPEG, GIF, WebP or AVIF under 10MB. The stored content type is normalised — uploads are not permitted to choose an arbitrary type, and responses carry X-Content-Type-Options: nosniff.",
        "operationId": "uploadPlaylistArtwork",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file. Field name read via formData.get(\"file\")."
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Playlist updated with the new artworkKey.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playlist": {
                      "$ref": "#/components/schemas/Playlist"
                    }
                  },
                  "required": [
                    "playlist"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing file.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not owned by the caller, or does not exist (non-enumerable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Artwork exceeds 10MB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/comments/playlist/{playlistId}": {
      "get": {
        "summary": "List playlist-level comments",
        "description": "Gated: owner session or a valid share token, via Bearer or `?token=`. Only returns comments with no trackId (playlist-level, not track-anchored). Deleted-field stripped; threaded with replies.",
        "operationId": "listPlaylistComments",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "tokenQuery": []
          }
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Threaded comments.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ThreadedComment"
                      }
                    }
                  },
                  "required": [
                    "comments"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Playlist private, nonexistent, or inaccessible (non-enumerable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}